requirements.txt
PyQt6>=6.7.0
azure-cosmos==4.5.1
requests==2.31.0
azure-identity==1.15.0
certifi>=2023.7.22
cairosvg==2.6.0
Pillow==10.4.0
